Split package into core and server variants
