+++ packages/nql/index.js
@@ -1,2 +1,2 @@
 // Export the Public API from lib
-module.exports = require('./lib/nql');
+module.exports = require('./lib/server');
