+++ packages/nql/test/unit/core.test.js
@@ -1,0 +1,23 @@
+require('../utils');
+
+const nqlCore = require('../../lib/core');
+const sandbox = sinon.createSandbox();
+
+describe('Core API', function () {
+    afterEach(function () {
+        sandbox.restore();
+    });
+
+    describe('querySQL', function () {
+        it('is not implemented', function () {
+            const query = nqlCore('id:3');
+
+            try {
+                query.querySQL();
+                should.fail();
+            } catch (err) {
+                err.message.should.eql('querySQL is not implemented in the browser');
+            }
+        });
+    });
+});
